Add the KVStore implementation to the bleve registry
I think as written this will not work; the config is supposed to be
json-marshallable.

Additionally, bleve seems to *really* want to store stuff on the
filesystem; the docs suggest that even providing our own storage backend
it will still read some stuff from disk. Perhaps we should reconsider
the approach.